refactor: separate extraction from construction
